Preparing firmware for write
Added TX EUSART for debugging purposes (will power board with DC to test UI and buttons)

Changed logic ISR timer to being 100 ms and count, that way we can use this timer for multiple purposes.

Starting to framework out the variables and functions we will need to run the heater.

Trying to comment as much as possible as I go to prevent needing to come back later and tackle documentation.