CHIP CHANGE
Got working NTC calculations and also had to make a change to the schematic (for clarity on which pin was high and low on the NTC).

Once this was working I discovered i had 99% program usage even with custom log functions to save space. I moved this project to PIC16F1847 which has double the program memory which I will need.